Extract ZipFileWorker in its own file.
Split the generate.js file into generate/index.js and
generate/ZipFileWorker.js.
