+++ lib/generate/ZipFileWorker.js
@@ -1,12 +1,10 @@
 'use strict';
 
-var utils = require('./utils');
-var GenericWorker = require('./stream/GenericWorker');
-var utf8 = require('./utf8');
-var crc32 = require('./crc32');
-var signature = require('./signature');
-var compressions = require('./compressions');
-
+var utils = require('../utils');
+var GenericWorker = require('../stream/GenericWorker');
+var utf8 = require('../utf8');
+var crc32 = require('../crc32');
+var signature = require('../signature');
 
 /**
  * Transform an integer into a string in hexadecimal.
@@ -304,75 +302,6 @@ var generateDataDescriptors = function (streamInfo) {
     return descriptor;
 };
 
-/**
- * Find the compression to use.
- * @param {String} fileCompression the compression defined at the file level, if any.
- * @param {String} zipCompression the compression defined at the load() level.
- * @return {Object} the compression object to use.
- */
-var getCompression = function (fileCompression, zipCompression) {
-
-    var compressionName = fileCompression || zipCompression;
-    var compression = compressions[compressionName];
-    if (!compression) {
-        throw new Error(compressionName + " is not a valid compression method !");
-    }
-    return compression;
-};
-
-/**
- * Create a worker to generate a zip file.
- * @param {Object} files the hash containing the ZipObjects.
- * @param {Object} options to generate the zip file.
- * @param {String} comment the comment to use.
- */
-exports.generateWorker = function (files, options, comment) {
-
-    var zipFileWorker = new ZipFileWorker(options.streamFiles, comment, options.platform);
-    var entriesCount = 0;
-    try {
-        for (var name in files) {
-            if (!files.hasOwnProperty(name)) {
-                continue;
-            }
-            entriesCount++;
-            var file = files[name];
-            var compression = getCompression(file.options.compression, options.compression);
-            var compressionOptions = file.options.compressionOptions || options.compressionOptions || {};
-            var dir, date;
-
-            // handle the deprecated options.dir
-            if (file._initialMetadata.dir !== file.dir) {
-                dir = file.dir;
-            } else {
-                dir = file.options.dir;
-            }
-
-            // handle the deprecated options.date
-            if(file._initialMetadata.date !== file.date) {
-                date = file.date;
-            } else {
-                date = file.options.date;
-            }
-
-            file._compressWorker(compression, compressionOptions)
-            .withStreamInfo("file", {
-                name : file.name,
-                dir : dir,
-                date : date,
-                comment : file.comment || "",
-                unixPermissions : file.unixPermissions,
-                dosPermissions : file.dosPermissions
-            })
-            .pipe(zipFileWorker);
-        }
-        zipFileWorker.entriesCount = entriesCount;
-    } catch (e) {
-        zipFileWorker.error(e);
-    }
-
-    return zipFileWorker;
-};
 
 /**
  * A worker to concatenate other workers to create a zip file.
@@ -579,3 +508,5 @@ ZipFileWorker.prototype.cleanUp = function () {
     }
     GenericWorker.prototype.cleanUp.call(this);
 };
+
+module.exports = ZipFileWorker;
