+++ lib/generate/index.js
@@ -1,0 +1,74 @@
+'use strict';
+
+var compressions = require('../compressions');
+var ZipFileWorker = require('./ZipFileWorker');
+
+/**
+ * Find the compression to use.
+ * @param {String} fileCompression the compression defined at the file level, if any.
+ * @param {String} zipCompression the compression defined at the load() level.
+ * @return {Object} the compression object to use.
+ */
+var getCompression = function (fileCompression, zipCompression) {
+
+    var compressionName = fileCompression || zipCompression;
+    var compression = compressions[compressionName];
+    if (!compression) {
+        throw new Error(compressionName + " is not a valid compression method !");
+    }
+    return compression;
+};
+
+/**
+ * Create a worker to generate a zip file.
+ * @param {Object} files the hash containing the ZipObjects.
+ * @param {Object} options to generate the zip file.
+ * @param {String} comment the comment to use.
+ */
+exports.generateWorker = function (files, options, comment) {
+
+    var zipFileWorker = new ZipFileWorker(options.streamFiles, comment, options.platform);
+    var entriesCount = 0;
+    try {
+        for (var name in files) {
+            if (!files.hasOwnProperty(name)) {
+                continue;
+            }
+            entriesCount++;
+            var file = files[name];
+            var compression = getCompression(file.options.compression, options.compression);
+            var compressionOptions = file.options.compressionOptions || options.compressionOptions || {};
+            var dir, date;
+
+            // handle the deprecated options.dir
+            if (file._initialMetadata.dir !== file.dir) {
+                dir = file.dir;
+            } else {
+                dir = file.options.dir;
+            }
+
+            // handle the deprecated options.date
+            if(file._initialMetadata.date !== file.date) {
+                date = file.date;
+            } else {
+                date = file.options.date;
+            }
+
+            file._compressWorker(compression, compressionOptions)
+            .withStreamInfo("file", {
+                name : file.name,
+                dir : dir,
+                date : date,
+                comment : file.comment || "",
+                unixPermissions : file.unixPermissions,
+                dosPermissions : file.dosPermissions
+            })
+            .pipe(zipFileWorker);
+        }
+        zipFileWorker.entriesCount = entriesCount;
+    } catch (e) {
+        zipFileWorker.error(e);
+    }
+
+    return zipFileWorker;
+};
